replaced all into with long
regexp replace (?<=\W)(Int)(?=\W)
by Long

replaced all int literals with long literals

regexp replace (?<![\.a-zA-Z0-9_])[0-9]+(?!\.)
by $0L

issues with standard API and, as expected: arrays

+++ oscar-cbls/src/main/scala/oscar/cbls/algo/dag/ConcreteDAG.scala
@@ -25,14 +25,14 @@ package oscar.cbls.algo.dag
   * @author renaud.delandtsheer@cetic.be
   * @param _UniqueID: an ID to be used as comparison for storage in sorted data structures
   */
-class ConcreteDAGNode(val _UniqueID:Int) extends DAGNode{
+class ConcreteDAGNode(val _UniqueID:Long) extends DAGNode{
 
   uniqueID = _UniqueID
 
   var PrecedingNodes: List[DAGNode] = List.empty
   var SucceedingNodes:List[DAGNode] = List.empty
 
-  final def compare(that: DAGNode):Int = {
+  final def compare(that: DAGNode):Long = {
     assert(this.uniqueID != that.uniqueID || this == that)
     this.uniqueID - that.uniqueID
   }
