+++ oscar-cbls/src/main/scala/oscar/cbls/algo/fun/PiecewiseLinearFunNaive.scala
@@ -17,14 +17,14 @@ package oscar.cbls.algo.fun
  */
 
 sealed abstract class PiecewiseLinearFunNaive{
-  def apply(value:Int):Int
-  def updateBefore(fromIncuded:Int,toIncluded:Int,update:LinearTransform):PiecewiseLinearFunNaive =
+  def apply(value:Long):Long
+  def updateBefore(fromIncuded:Long,toIncluded:Long,update:LinearTransform):PiecewiseLinearFunNaive =
     new UpdatedPiecewiseLinearFunNaive(fromIncuded,toIncluded,update:LinearTransform,this)
 }
 case object IdentityNaive extends PiecewiseLinearFunNaive{
-  override def apply(value:Int):Int = value
+  override def apply(value:Long):Long = value
 }
-case class UpdatedPiecewiseLinearFunNaive(fromIncuded:Int,toIncluded:Int,update:LinearTransform,base:PiecewiseLinearFunNaive) extends PiecewiseLinearFunNaive{
-  override def apply(value:Int):Int = if(value >= fromIncuded && value <= toIncluded) base(update(value)) else base(value)
+case class UpdatedPiecewiseLinearFunNaive(fromIncuded:Long,toIncluded:Long,update:LinearTransform,base:PiecewiseLinearFunNaive) extends PiecewiseLinearFunNaive{
+  override def apply(value:Long):Long = if(value >= fromIncuded && value <= toIncluded) base(update(value)) else base(value)
 }
